requirements.txt
Django==4.0.2
djangorestframework==3.13.1
mysqlclient==2.1.0
sqlparse==0.4.2
bcrypt==3.2.0
boto3==1.21.16
pillow==9.0.1
python-statsd==2.1.0
mod_wsgi==4.9.0